more updates

+++ src/main/java/com/mrose/financial/MapSummation.java
@@ -1,18 +1,17 @@
 package com.mrose.financial;
 
 import com.google.common.base.Function;
-
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Map;
 
 /**
  * Created by martinrose on 10/29/14.
  */
-class MapSummation<K> implements Function<Map<K, BigDecimal>, BigDecimal>  {
+class MapSummation<K> implements Function<Map<K, BigInteger>, BigInteger>  {
   @Override
-  public BigDecimal apply(Map<K, BigDecimal> input) {
-    BigDecimal running = BigDecimal.ZERO;
-    for( BigDecimal bd: input.values()) {
+  public BigInteger apply(Map<K, BigInteger> input) {
+    BigInteger running = BigInteger.ZERO;
+    for( BigInteger bd: input.values()) {
       running = running.add(bd);
     }
     return running;
